Scraper is indeed working
Using requests and BeautifulSoup to POST to LeopardWeb and get session tokens to allow us to post to the Advanced Search and select all classes. This returns a HUGE HTML page with every class. We sift through this and turn it into a large JSON array and save it to a file.

qs file: A qs file is simply a file that stores the magic query string when posting to the Advanced Search. It's a huge form with many parameters we need to automate. This will likely change in a future implementation that generates this on the fly. The API will most likely not change, but implementation will.

More to come.