Working commit
The decorators from 81359250926ff32be818194a813419f1dfe2948f caused a problem with nesting and we are unable to use them as planned. This isn't a big problem, but it will lead to possible inconsistencies in error checking. Maybe we can find a way around it, but it may not involve decorators. They have been removed and things should be working as planned.