Fixed an issue where some classes would have most columns empty.
This is when classes may have different times during the week. Such as M and W are 8-8:50, but F is 8-9:15. This looks as such:

COMP 1070 Computer Science I MW 8am-8:50am
                              F 8am-9:15am

We easily handle this by keeping track of the previous and using specific columns to detect when we've found an empty one. After detection, we fill in the missing data with a new function from utils: dict_safe_update. This function is similar to jQuery.extend; it merges two dicts together without overwriting properties on the old one (needs unit testing). This should actually be refactored to be done in the scrape function, instead of cleanup_data.

Discovered an issue when a class may have multiple professors. This doesn't happen often, but there are some in the current database. This needs to be handled and tested. Splitting on comma, then the parentheses strip for each should work just fine.

Also handled cases where some entries would have just whitespace in some columns. The attribute column and empty entries (explained above) are the most guilty. We handle this by simply calling .strip() on each column before throwing it in the data array.