Deprecated the connect(self) function. It was useless.
All it did was open the requests.Session object. We can easily handle this in the constructor without requiring the user of the API properly opens the session. This still makes us require checks in each function for authentication properly, which we don't do. Perhaps look at the decorators from an earlier commit. They gave us some trouble with nesting, but might come in handy for a single purpose.