Implemented authenticators to handle session logic
This drastically rips out a part of the scraper that shouldn't be handled by the scraper itself.
The scraper should only be concerned with scraping data, not handling sessions.
So we separated this logic into the authenticators classes. In there, sessions are created, posted to, and used to grab data all under the session token and cookies stored from the previous.

NOTE: DO NOT USE THE _session VARIABLE IN THE AUTHENTICATOR.
The useful methods from _session should be properly exposed as a middleman for error handling or packet manipulation (as we do with post, get, and put). It may not be useful for this particular scraper, but at least we can abstract that functionality.